fix(wresp): 删除 C# 兼容 length 模块；修 i32 边界长度取负的 debug 溢出
- 删除 src/length.rs（RespLengthEncodingUtils 转写）：全仓无任何使用方，
  且其 32 位分支刻意复刻了 C# TryReadLength 读 input[0..4]（含信号字节）
  的上游 bug——任务书明确不需要兼容 C#，此类 hack 一并清理
- read.rs try_read_signed_length_header：负长度经 i64 取负后再 as i32。
  原写法 -(value as i32) 在 value == 2^31（$-2147483648\r\n 头）时
  as i32 得 i32::MIN，对其取负在 debug/测试构建触发
  'attempt to negate with overflow' panic；release 下静默回绕仅因巧合正确
- 补 i32 边界（正最大/负最小/正溢出）回归测试

+++ node/wresp/src/read.rs
@@ -280,9 +280,10 @@ pub fn try_read_signed_length_header(
     });
   }
 
-  // Convert to signed value
+  // Convert to signed value（经 i64 取负：value == 2^31 时 as i32 即 i32::MIN，
+  // 若在 i32 域直接取负 debug 构建会溢出 panic）
   *length = if negative {
-    -(value as i32)
+    -(value as i64) as i32
   } else {
     value as i32
   };
